requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.22.10
botocore==1.25.10
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Flask==2.1.2
grpcio==1.46.0
grpcio-tools==1.46.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.29
protobuf==3.20.1
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
validators==0.19.0
vine==5.0.0
Wave==0.0.2
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
